tazama-rule-agent/requirements.txt
streamlit>=1.35.0
openai>=1.30.0
python-dotenv>=1.0.0
rich>=13.0.0
pexpect>=4.9.0
